+++ src/comp_types/hashed_comp.hpp
@@ -34,7 +34,7 @@ public:
   HashedComp(const HashedComp&) = default;
   HashedComp& operator=(const HashedComp&) = default;
   inline HashedComp(void* hash_seed, const Comp& r_comp);
-  uint32_t bignum_bytes() const { return BaseComp::bignum_bytes(); }
+  uint64_t bignum_bytes() const { return BaseComp::bignum_bytes(); }
 
   uint64_t get_clhashkey() const { return BaseComp::clhashkey_; }
   bool equals_clhashkey(const HashedComp &comp) const {
